requirements.txt
python-dotenv
huggingface_hub
langchain
langchain-ollama
streamlit
langchain_community
langchain-huggingface
streamlit-chat 
PyPDF2
faiss-gpu
faiss-cpu